Renamed folders from "export_plugin" to "shrinkwrap", renamed cmake flag BUILD_CVMFS_SHRINKWRAP to BUILD_SHRINKWRAP, minor code improvements (in paticular linting errors)
