+++ ci/build_preloader.sh
@@ -26,6 +26,6 @@ cmake -DBUILD_PRELOADER=on \
   -DBUILD_SERVER=no \
   -DBUILD_LIBCVMFS=no \
   -DBUILD_LIBCVMFS_CACHE=no \
-  -DBUILD_CVMFS_SHRINKWRAP=no
+  -DBUILD_SHRINKWRAP=no
   $REPO_ROOT
 make -j $(nproc)
