+++ cmake/Modules/cvmfs_options.cmake
@@ -12,7 +12,7 @@ else (MACOSX)
 endif(MACOSX)
 option (BUILD_LIBCVMFS          "Build the CernVM-FS client library"                               ON)
 option (BUILD_LIBCVMFS_CACHE    "Build the cache plugin library"                                   ON)
-option (BUILD_CVMFS_SHRINKWRAP  "Build the shrinkwrap tool"                                        OFF)
+option (BUILD_SHRINKWRAP  "Build the shrinkwrap tool"                                        OFF)
 
 option (BUILD_PRELOADER         "Build the alien cache pre-loader"                                 OFF)
 option (BUILD_OCTOPUS           "Build the octopus REST server"                                    OFF)
@@ -42,7 +42,7 @@ if (BUILD_ALL)
     set (BUILD_SERVER ON)
     set (BUILD_LIBCVMFS ON)
     set (BUILD_LIBCVMFS_CACHE ON)
-    set (BUILD_CVMFS_SHRINKWRAP ON)
+    set (BUILD_SHRINKWRAP ON)
     set (BUILD_PRELOADER ON)
     set (BUILD_OCTOPUS ON)
   endif (!MACOSX)
