--- cvmfs/export_plugin/fs_traversal_libcvmfs.h
@@ -1,9 +1,0 @@
-/**
- * This file is part of the CernVM File System.
- */
-#ifndef CVMFS_EXPORT_PLUGIN_FS_TRAVERSAL_LIBCVMFS_H_
-#define CVMFS_EXPORT_PLUGIN_FS_TRAVERSAL_LIBCVMFS_H_
-
-struct fs_traversal *libcvmfs_get_interface();
-
-#endif  // CVMFS_EXPORT_PLUGIN_FS_TRAVERSAL_LIBCVMFS_H_
