--- cvmfs/export_plugin/posix/interface.h
@@ -1,9 +1,0 @@
-/**
- * This file is part of the CernVM File System.
- */
-#ifndef CVMFS_EXPORT_PLUGIN_POSIX_INTERFACE_H_
-#define CVMFS_EXPORT_PLUGIN_POSIX_INTERFACE_H_
-
-struct fs_traversal *posix_get_interface();
-
-#endif  // CVMFS_EXPORT_PLUGIN_POSIX_INTERFACE_H_
