+++ cvmfs/shrinkwrap/fs_traversal.cc
@@ -13,13 +13,13 @@
 
 #include "atomic.h"
 #include "cvmfs_config.h"
-#include "export_plugin/fs_traversal.h"
-#include "export_plugin/fs_traversal_interface.h"
-#include "export_plugin/fs_traversal_libcvmfs.h"
-#include "export_plugin/posix/interface.h"
-#include "export_plugin/spec_tree.h"
 #include "libcvmfs.h"
 #include "logging.h"
+#include "shrinkwrap/fs_traversal.h"
+#include "shrinkwrap/fs_traversal_interface.h"
+#include "shrinkwrap/fs_traversal_libcvmfs.h"
+#include "shrinkwrap/posix/interface.h"
+#include "shrinkwrap/spec_tree.h"
 #include "smalloc.h"
 #include "statistics.h"
 #include "util/posix.h"
