+++ cvmfs/shrinkwrap/fs_traversal_libcvmfs.cc
