+++ cvmfs/shrinkwrap/fs_traversal_libcvmfs.h
@@ -1,0 +1,9 @@
+/**
+ * This file is part of the CernVM File System.
+ */
+#ifndef CVMFS_SHRINKWRAP_FS_TRAVERSAL_LIBCVMFS_H_
+#define CVMFS_SHRINKWRAP_FS_TRAVERSAL_LIBCVMFS_H_
+
+struct fs_traversal *libcvmfs_get_interface();
+
+#endif  // CVMFS_SHRINKWRAP_FS_TRAVERSAL_LIBCVMFS_H_
