+++ cvmfs/shrinkwrap/posix/garbage_collector.h
@@ -1,8 +1,8 @@
 /**
  * This file is part of the CernVM File System.
  */
-#ifndef CVMFS_EXPORT_PLUGIN_POSIX_GARBAGE_COLLECTOR_H_
-#define CVMFS_EXPORT_PLUGIN_POSIX_GARBAGE_COLLECTOR_H_
+#ifndef CVMFS_SHRINKWRAP_POSIX_GARBAGE_COLLECTOR_H_
+#define CVMFS_SHRINKWRAP_POSIX_GARBAGE_COLLECTOR_H_
 
 #include "statistics.h"
 
@@ -17,4 +17,4 @@ void InitializeGarbageCollection(struct fs_traversal_context *ctx);
 void FinalizeGarbageCollection(struct fs_traversal_context *ctx);
 int RunGarbageCollection(struct fs_traversal_context *ctx);
 
-#endif  // CVMFS_EXPORT_PLUGIN_POSIX_GARBAGE_COLLECTOR_H_
+#endif  // CVMFS_SHRINKWRAP_POSIX_GARBAGE_COLLECTOR_H_
