+++ cvmfs/shrinkwrap/posix/interface.cc
@@ -14,13 +14,13 @@
 #include <map>
 #include <string>
 
-#include "export_plugin/fs_traversal_interface.h"
-#include "export_plugin/util.h"
 #include "garbage_collector.h"
 #include "hash.h"
 #include "helpers.h"
 #include "libcvmfs.h"
 #include "logging.h"
+#include "shrinkwrap/fs_traversal_interface.h"
+#include "shrinkwrap/util.h"
 #include "smalloc.h"
 #include "string.h"
 #include "util/posix.h"
