+++ cvmfs/shrinkwrap/posix/interface.h
@@ -1,0 +1,9 @@
+/**
+ * This file is part of the CernVM File System.
+ */
+#ifndef CVMFS_SHRINKWRAP_POSIX_INTERFACE_H_
+#define CVMFS_SHRINKWRAP_POSIX_INTERFACE_H_
+
+struct fs_traversal *posix_get_interface();
+
+#endif  // CVMFS_SHRINKWRAP_POSIX_INTERFACE_H_
