+++ cvmfs/shrinkwrap/scripts/build.sh
@@ -1,6 +1,7 @@
 #
 # This file is part of the CernVM File System.
 #
+
 print_help() {
   echo "build.sh --spec_dir=<directory with spec files> --cvmfs_config=<cvmfs config> [--dest_sys=<destination file system>] [--dest_base=<destination file system base>] [--dest_data=<destination data directory>] [--gc]"
 }
