+++ cvmfs/shrinkwrap/scripts/docker_inject/setup.sh
@@ -1,4 +1,5 @@
 #
 # This file is part of the CernVM File System.
 #
+
 pip install --user python-dxf
