+++ cvmfs/shrinkwrap/scripts/spec_builder.py
