+++ cvmfs/shrinkwrap/scripts/spec_test.py
