+++ cvmfs/shrinkwrap/scripts/trace.sh
@@ -1,6 +1,7 @@
 #
 # This file is part of the CernVM File System.
 #
+
 TRACE_DIR=/tmp/cvmfs-trace
 SPEC_DIR=/tmp/cvmfs-spec
 POLICY="pdir"
