+++ cvmfs/shrinkwrap/spec_tree.cc
