+++ cvmfs/shrinkwrap/util.cc
